Reworked core module structure
